# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/me": {
      "get": {
        "operationId": "UserController_getMe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateMe_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Individual account deletion per spec/behavior/data-retention.md: profile media is purged first, PII is scrubbed to a \"Deleted User\" tombstone, historical records stay preserved anonymized, the analytics forget is confirmed, and the Supabase Auth account is deleted last. A 502 means the flow did not finish — depending on the failing step the account may already be anonymized (sign-in still works until the final step succeeds) — and every step is idempotent, so simply retry until it returns success.",
        "operationId": "UserController_deleteMe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete current account (irreversible)",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/me/permissions": {
      "get": {
        "description": "Returns the caller's flattened permission set for the chapter identified by the `x-chapter-id` header. Clients use this to render permission-aware UI without duplicating RBAC rules or issuing one request per role.",
        "operationId": "UserController_getMyPermissions_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyPermissionsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get effective permissions for the active chapter",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/me/avatar-url": {
      "post": {
        "operationId": "UserController_requestAvatarUploadUrl_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAvatarUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get signed upload URL for profile photo",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Frapp API",
    "description": "The Operating System for Greek Life",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "chapter-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-chapter-id"
      }
    },
    "schemas": {
      "MyPermissionsDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "description": "Caller's effective permission set for the active chapter. Contains the wildcard `*` for Presidents. Empty array means no active roles.",
            "example": [
              "members:view",
              "events:create"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string"
          },
          "graduation_year": {
            "type": "number",
            "nullable": true
          },
          "current_city": {
            "type": "string"
          },
          "current_company": {
            "type": "string"
          }
        }
      },
      "RequestAvatarUploadUrlDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Original filename for the avatar image"
          },
          "content_type": {
            "type": "string",
            "description": "MIME content type (e.g. image/jpeg, image/png)"
          }
        },
        "required": [
          "filename",
          "content_type"
        ]
      },
...
    }
  }
}